Move family header data into own util and update drawer and hook
